PuzzleFunctions created. reset and onChange functions can be abstracted away
